move around helper functions to import utils from timeseries
