+++ darts/tests/dataprocessing/encoders/test_encoders.py
@@ -30,6 +30,7 @@ from darts.dataprocessing.encoders.encoders import (
 from darts.dataprocessing.transformers import Scaler
 from darts.logging import get_logger, raise_log
 from darts.utils import timeseries_generation as tg
+from darts.utils.utils import generate_index
 
 logger = get_logger(__name__)
 
@@ -79,7 +80,7 @@ class TestEncoder:
     # multi-TS at prediction should be as follows
     inf_ts_short_future = [
         TimeSeries.from_times_and_values(
-            tg.generate_index(
+            generate_index(
                 start=ts.end_time() + (1 - 12) * ts.freq, length=12 + 6, freq=ts.freq
             ),
             np.arange(12 + 6),
@@ -89,7 +90,7 @@ class TestEncoder:
 
     inf_ts_long_future = [
         TimeSeries.from_times_and_values(
-            tg.generate_index(
+            generate_index(
                 start=ts.end_time() + (1 - 12) * ts.freq, length=12 + 8, freq=ts.freq
             ),
             np.arange(12 + 8),
@@ -99,7 +100,7 @@ class TestEncoder:
 
     inf_ts_short_past = [
         TimeSeries.from_times_and_values(
-            tg.generate_index(
+            generate_index(
                 start=ts.end_time() + (1 - 12) * ts.freq, length=12, freq=ts.freq
             ),
             np.arange(12),
@@ -109,7 +110,7 @@ class TestEncoder:
 
     inf_ts_long_past = [
         TimeSeries.from_times_and_values(
-            tg.generate_index(
+            generate_index(
                 start=ts.end_time() + (1 - 12) * ts.freq,
                 length=12 + (8 - 6),
                 freq=ts.freq,
@@ -667,7 +668,7 @@ class TestEncoder:
 
         attribute = "month"
         month_series = TimeSeries.from_times_and_values(
-            times=tg.generate_index(
+            times=generate_index(
                 start=pd.to_datetime("2000-01-01"), length=24, freq="MS"
             ),
             values=np.arange(24),
@@ -724,7 +725,7 @@ class TestEncoder:
         attribute = "month"
 
         month_series = TimeSeries.from_times_and_values(
-            times=tg.generate_index(
+            times=generate_index(
                 start=pd.to_datetime("2000-01-01"), length=24, freq="MS"
             ),
             values=np.arange(24),
@@ -930,7 +931,7 @@ class TestEncoder:
 
         # inference set
         pc, fc = encs.encode_inference(n=12, target=ts)
-        year_index = tg.generate_index(
+        year_index = generate_index(
             start=ts.end_time() - ts.freq * (input_chunk_length - 1),
             length=24,
             freq=ts.freq,
