+++ darts/utils/__init__.py
@@ -7,5 +7,5 @@ from .utils import (
     _build_tqdm_iterator,
     _parallel_apply,
     _with_sanity_checks,
-    retain_period_common_to_all,
+    n_steps_between,
 )
