+++ darts/utils/historical_forecasts/optimized_historical_forecasts_torch.py
@@ -16,7 +16,7 @@ from darts.utils.historical_forecasts.utils import (
     _get_historical_forecast_boundaries,
     _process_predict_start_points_bounds,
 )
-from darts.utils.timeseries_generation import generate_index
+from darts.utils.utils import generate_index
 
 logger = get_logger(__name__)
